Support PHP version mangers
Sorry for the bad PR, I don't know TypeScript.

This PR sets the current working directory of the child process so PHP version managers work correctly.

By default, the child process current working directory is `process.cwd()`, which might not always be in the project tree. For PHP version managers, `php` needs to be executed inside the project tree so the `.phpversion` (or whatever similar file) is detected correctly. The solution is to set the child process cwd to `path.dirname(vscode.window.activeTextEditor.document.fileName)`.

 For example, here is my `mampenv` package that allows a `.phpversion` file to be specified in the root of the project, much like `rbenv` etc. https://github.com/benallfree/mampenv

Hope it's clear :)